Add files via upload
This project is done for a company called "WeFit" which have three child companies "FitWear", "LeanFit" and  "SleepyNights" this project focused on finding potential customers for "WeFit", "FitWear" and "SleepyNights" based on the existing dataset of "WeFit". This project also gives some valuable insights regarding the sales of WeFit and This project also Builds a sales dashboard for WeFit. Entire project is done with the help of MS Excel.